SlidesTable refactored
